Implement authentication for meeting and registration Controller

+++ app/Http/Controllers/MeetingController.php
@@ -6,12 +6,15 @@ use Illuminate\Http\Request;
 use App\Meeting;
 use Carbon\Carbon;
 use App\User;
+use Tymon\JWTAuth\Facades\JWTAuth;
 
 class MeetingController extends Controller
 {
     public function __construct()
     {
-        // $this->>middleware('name');
+        $this->middleware('jwt.auth', ['only'=>[
+            'update', 'store', 'destroy'
+        ]]);
     }
 
     /**
@@ -48,14 +51,17 @@ class MeetingController extends Controller
         $this->validate($request, [
             'title' => 'required',
             'description' => 'required',
-            'time' => 'required|date_format:YmdHie',
-            'user_id' => 'required'
+            'time' => 'required|date_format:YmdHie'
         ]);
+
+        if(! $user = JWTAuth::parseToken()-> authenticate()){
+            return response()->json(['msg' => 'user not found'], 404);
+        }
 
         $title = $request->input('title');
         $description = $request->input('description');
         $time = $request->input('time');
-        $user_id = $request->input('user_id');
+        $user_id = $user->id;
 
         $meeting = new Meeting([
             'time' => Carbon::createFromFormat('YmdHie', $time),
@@ -115,14 +121,17 @@ class MeetingController extends Controller
         $this->validate($request, [
             'title' => 'required',
             'description' => 'required',
-            'time' => 'required|date_format:YmdHie',
-            'user_id' => 'required'
+            'time' => 'required|date_format:YmdHie'
         ]);
+
+        if(! $user = JWTAuth::parseToken()-> authenticate()){
+            return response()->json(['msg' => 'user not found'], 404);
+        }
 
         $title = $request->input('title');
         $description = $request->input('description');
         $time = $request->input('time');
-        $user_id = $request->input('user_id');
+        $user_id = $user->id;
 
         $meeting = Meeting::with('users')->findOrFail($id);
 
@@ -158,6 +167,12 @@ class MeetingController extends Controller
     public function destroy($id)
     {
         $meeting = Meeting::findOrFail($id);
+        if(! $user = JWTAuth::parseToken()-> authenticate()){
+            return response()->json(['msg' => 'user not found'], 404);
+        }
+        if (!$meeting->users()->where('users.id', $user->id)->first()) {
+            return response()->json(['msg' => 'user not registered for meeting, update not successful'], 401);
+        };
         $users = $meeting->users;
         $meeting->users()->detach();
         if (!$meeting->delete()) {
